Move arrow-json-integration-test to new crate
